# src/main/java/ie/kevmurf/oas/api/StudentsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.21-SNAPSHOT).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ie.kevmurf.oas.api;

import ie.kevmurf.oas.model.StudentApiSpec;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-07-08T20:12:16.400+01:00[Europe/London]")
@Api(value = "students", description = "the students API")
public interface StudentsApi {

    @ApiOperation(value = "Creates a new student.", nickname = "createStudent", notes = "", response = StudentApiSpec.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Created", response = StudentApiSpec.class) })
    @RequestMapping(value = "/students",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<StudentApiSpec> createStudent(@ApiParam(value = ""  )  @Valid @RequestBody StudentApiSpec body
);


    @ApiOperation(value = "Delete a student.", nickname = "deleteStudent", notes = "", tags={  })
    @ApiResponses(value = {  })
    @RequestMapping(value = "/students/{studentId}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteStudent(@Min(1)@ApiParam(value = "The ID of the student to delete.",required=true, allowableValues="") @PathVariable("studentId") Integer studentId
);


    @ApiOperation(value = "Returns a student by ID.", nickname = "getStudentById", notes = "", response = StudentApiSpec.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = StudentApiSpec.class),
        @ApiResponse(code = 400, message = "The specified student ID is invalid (e.g. not a number)."),
        @ApiResponse(code = 404, message = "A student with the specified ID was not found."),
        @ApiResponse(code = 200, message = "Unexpected error") })
    @RequestMapping(value = "/students/{studentId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<StudentApiSpec> getStudentById(@Min(1)@ApiParam(value = "The ID of the student to retrieve.",required=true, allowableValues="") @PathVariable("studentId") Integer studentId
);


    @ApiOperation(value = "Returns a list of students.", nickname = "getStudents", notes = "", response = StudentApiSpec.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = StudentApiSpec.class, responseContainer = "List") })
    @RequestMapping(value = "/students",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<StudentApiSpec>> getStudents();


    @ApiOperation(value = "Update / Replace a student.", nickname = "updateStudent", notes = "", tags={  })
    @ApiResponses(value = {  })
    @RequestMapping(value = "/students/{studentId}",
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Void> updateStudent(@Min(1)@ApiParam(value = "The ID of the student to update.",required=true, allowableValues="") @PathVariable("studentId") Integer studentId
,@ApiParam(value = ""  )  @Valid @RequestBody StudentApiSpec body
);

}